Add source and copyright notice for CIS files from pcmcia-cs
The CIS files mostly come from pcmcia-cs releases, in which the source
files were named etc/cis/*.cis and the binaries were named
etc/cis/*.dat.  However, the current drivers request the *binaries* as
cis/*.cis.  It's too late to fix that, so put the sources in a
subdirectory here.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>