rt2860sta: Use latest firmware for RT3090
Ralink's original drivers for RT2800P/RT2800E-family chips used
multiple different versions of the firmware for different chips.  The
rt2860sta driver in staging was briefly converted to load different
files for different chips.  However, the current rt2860.bin is
supposed to work on all of them, so replace rt3090.bin with a symlink.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>