cis: Add makefile for building CIS files from text format
Most of the CIS files come from the pcmcia-cs project and are built
from a text format using the pack_cis tool.

In pcmcia-cs 3.2.8, apparently the last release, pack_cis had some
bugs that are fatal when it is built with a current gcc and glibc.
Therefore I refer to my own repackaged version that runs successfully
and generates exactly the same binaries included here.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>