linux-firmware: brcm: remove nvram file for bcm4329 development board
The nvram file contains configuration info for firmware which varies
with different hardware designs. The version currently in the
repository is for a Broadcom-internal development board that is not
available on the market. This file is not applicable for any exsisting
end-user product. Users should use the nvram file that was
shipped with their device to avoid unexpected or incorrect behavior.
Remove the file from repository to avoid confusion.

Signed-off-by: Franky Lin <frankyl@broadcom.com>
Signed-off-by: Arend van Spriel <arend@broadcom.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>