Add go7007 firmware.
The go7007 staging driver has been overhauled and is now in decent shape. As
part of that work all related firmware files were collected into one place
(the /lib/firmware/go7007 directory) and the .hex firmwares were converted
into a binary format to make it easy to load for the kernel. It's now time
to add the firmware files here to get full support for this driver.

The original firmware came from Micronas. The initial driver package from
them can be found here:

http://nikosapi.org/software/WIS_Go7007/wis-go7007-linux-0.9.8.tar.bz2

The s2250 firmware is from Sensoray and was already part of linux-firmware.
That firmware has just been renamed.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>