linux-firmware: Clean up i915 by removing old skl dmc firmware.
The code points to the major version that is a link to the most
recent one, so we can remove the old firmware blobs.

For tests and validation purposes we will let few minor versions
available, but we need to let linux-firmware repo as
clean as possible, so let's remove old and unused ones.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>