ath10k: QCA988X hw2.0: add board.bin
In commit 40d14143b787 ("ath10k: add firmware 10.2.4.45 as firmware-4.bin") I
forgot to commit board.bin file and that's why it was not included in the
commit even I had added an entry to WHENCE. Add that now.

At the same time also add Version field for firmware-4.bin.

Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>