ti-connectivity: add firmware for CC2560(A) Bluetooth
This adds TIInit_6.2.31.bts and TIInit_6.6.15.bts for TI CC2560 and
CC2560A Bluetooth chips. These are found in LEGO MINDSTORMS EV3, which
has mainline kernel support now.

The versions listed are the ones given on the TI wiki[1] where the files
were obtained. The version in parentheses are the ones found in the
files themselves.

These files are redistributable under the same LICENCE.ti-connectivity that
is used by other TI firmwares.

[1]: http://processors.wiki.ti.com/index.php/CC256x_Downloads

Signed-off-by: David Lechner <david@lechnology.com>
Signed-off-by: Josh Boyer <jwboyer@kernel.org>